full

+++ lib/widget/taskWordToDo.dart
@@ -1,19 +1,23 @@
 // ignore: file_names
+import 'package:confirm_dialog/confirm_dialog.dart';
 import 'package:flutter/material.dart';
 
 class taskWordToDo extends StatelessWidget {
   taskWordToDo({
     super.key,
     this.item,
+    required this.index,
     required this.handleDelete,
   });
   dynamic item;
+  var index;
   final Function handleDelete;
   @override
   Widget build(BuildContext context) {
     return Container(
       decoration: BoxDecoration(
-          borderRadius: BorderRadius.circular(15), color: Colors.brown),
+          borderRadius: BorderRadius.circular(15),
+          color: index % 2 == 0 ? Colors.brown : Colors.grey),
       width: double.infinity,
       height: 70,
       margin: const EdgeInsets.only(bottom: 20),
@@ -30,8 +34,11 @@ class taskWordToDo extends StatelessWidget {
                   color: Colors.white),
             ),
             InkWell(
-              onTap: () {
-                handleDelete(item.id);
+              onTap: () async {
+                if (await confirm(context)) {
+                  return handleDelete(item.id);
+                }
+                return;
               },
               child: const Icon(
                 Icons.delete_forever,
